requirements.txt
# Core dependencies for the example workflow
# Install with: pip install -r requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.0.0
python-dotenv>=1.0.0
